services/ai/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
anthropic>=0.42.0
openai>=1.0.0
asyncpg>=0.30.0
pgvector>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
redis>=5.2.0
pypdf>=5.0.0
python-docx>=1.1.0
httpx>=0.27.0
slowapi==0.1.9
